handle large positionId (#122)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatainsightsapi/location/model/Location.kt
@@ -6,7 +6,7 @@ import kotlin.time.ExperimentalTime
 data class Location
 @OptIn(ExperimentalTime::class)
 constructor(
-  val positionId: Int,
+  val positionId: Long,
   val deviceId: Int?,
   val gpsDate: Instant? = null,
   val speed: Int? = null,
